+++ src/home/HomePage.jsx
@@ -1,21 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Router, Route, browserHistory} from 'react-router';
-import HomeMenuBar from '../MenuBar.jsx';
+import HomeMenuBar from './HomeMenuBar.jsx';
 import SignUpTabs from './SignUpTabs.jsx';
 import LoginPage from './form/LoginForm.jsx';
-import AboutUs from './AboutUs.jsx';
+import AboutUs from '../AboutUs.jsx';
 import styles from '../../public/css/login_signup.css';
-import { Provider } from 'react-redux';
-import store from '../redux/store.js';
-
 
 const HomeComp = function() {
     return (
         <div>
-            <Provider store={store}>
-                <HomeMenuBar homeordash='home'/>
-            </Provider>
+            <HomeMenuBar/>
             <Router history={browserHistory}>
                 <Route path="/home" component={AboutUs}/>
                 <Route path="/home/about" component={AboutUs}/>
